Refactor code in progressindicator.go
Move humanReadableSize to utils.go
Move copyWithProgress to io.go
